Move the interface into parent because those interfaces will be used as default.
