+++ osu.Game.Rulesets.Karaoke/Edit/Generator/Beatmaps/Pages/PageGeneratorConfig.cs
@@ -1,0 +1,1 @@
+﻿
